Chore/migrate to sqlite (#27)
* Исправил ошибку маппинга c UserEntity, добавил нормализацию для номера телефона в DTO при создании пользователя, добавил мапперов для перевода из ORM моделей в доменные объекты

* Добавил Value object для Points, mapper для преобразования пользователя в dto, отрефакторил логику регистрации, добавив валидацию вводимых значений Ф.И.О., исправил логику зачисления баллов

* Добавление недостающих зависимостей в pyproject.toml

* Удаление лишнего файла

* Добавил Makefile для группировки команд для инструментов, добавил CI-pipeline, через Github Action, используя Makefile, для улучшения качества кода и повышения DX

* Добавил недостающие инструменты в dev-зависимости необходимые для CI-pipeline-а

* Убрал бесполезную директиву type: ignore в config.py

* Использование официального uv Action-а от Astral для ускорения его установки

* feat/добавление архитектурного манифеста и описания workflow проекта

* docs/Изменил название workflow на то, которое автоматически распознаётся Github

* docs/Использовал markdown linter на ADR-ах

* docs/Добавил шаблоны для issue и pull request-ов

* chore: мигрировал систему с PosgreSQL на SQLite

+++ src/pybot/db/models/comment_answers.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import BigInteger, ForeignKey
+from sqlalchemy import ForeignKey, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ..base_class import Base
@@ -9,5 +9,5 @@ from ..base_class import Base
 class CommentAnswer(Base):
     __tablename__ = "comment_answers"
 
-    comment_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("comments.id"), primary_key=True)
-    answer_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("comments.id"), primary_key=True)
+    comment_id: Mapped[int] = mapped_column(Integer, ForeignKey("comments.id"), primary_key=True)
+    answer_id: Mapped[int] = mapped_column(Integer, ForeignKey("comments.id"), primary_key=True)
