+++ src/pybot/db/models/comments.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sqlalchemy import TIMESTAMP, BigInteger, ForeignKey, Text
+from sqlalchemy import TIMESTAMP, ForeignKey, Integer, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ..base_class import Base
@@ -17,8 +17,8 @@ if TYPE_CHECKING:
 class Comment(Base):
     __tablename__ = "comments"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id"), nullable=False)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
     content: Mapped[str] = mapped_column(Text, nullable=False)
     datetime: Mapped[str] = mapped_column(TIMESTAMP, nullable=False)
